Fix SIGABRT on VANC parsing failure
If the protocol-specific VANC parser fails to parse the packet, the
code which removes the entry from user_data was copying the wrong
number of bytes.  This would result in a double-free condition if
there were less than 3 entries left in the list.

Properly represent the size of the user_data struct as 24 bytes,
rather than passing to memmove() the size of the pointer (i.e. 8-bytes).

Note that this problem would only manifest if the packet was malformed
but the VANC checksum was valid.  This is because invalid VANC
checksums would cause the packet to be dropped and never reach this
stage.  In the field we saw this in a case where the CEA-708 CDP checksum
failed but the VANC checksum was valid.